+++ src/constants/index.ts
@@ -117,49 +117,8 @@ export const S3_BUCKET = {
     FIREFLY_LENS_MEDIA: 'firefly-lens-media',
 };
 
-export const MAX_CHAR_SIZE_PER_POST: Record<SocialSource, number> = {
-    [Source.Farcaster]: 1024,
-    [Source.Lens]: 5000,
-    [Source.Twitter]: 280,
-};
-export const MAX_CHAR_SIZE_VERIFY_PER_POST: Record<SocialSource, number> = {
-    [Source.Farcaster]: 1024,
-    [Source.Lens]: 5000,
-    [Source.Twitter]: 25000,
-};
-export const MAX_IMAGE_SIZE_PER_POST: Record<SocialSource, number> = {
-    [Source.Farcaster]: 2,
-    [Source.Lens]: 20,
-    [Source.Twitter]: 4,
-};
-export const MAX_VIDEO_SIZE_PER_POST: Record<SocialSource, number> = {
-    [Source.Farcaster]: 400 * 1024 * 1024, // 400MB
-    [Source.Lens]: 400 * 1024 * 1024, // 400MB
-    [Source.Twitter]: 75 * 1024 * 1024, // 75MB
-};
-
 // HTTP Cache headers
 export const CACHE_AGE_INDEFINITE_ON_DISK = 'public, s-maxage=31536000, max-age=31536000, must-revalidate';
-
-export const MAX_FILE_SIZE_PER_IMAGE: Record<SocialSource, number> = {
-    [Source.Twitter]: 5 * 1024 * 1024, // 5MB
-    [Source.Lens]: 100 * 1024 * 1024, // 100MB
-    [Source.Farcaster]: 100 * 1024 * 1024, // 100MB
-};
-
-export const MAX_FILE_SIZE_PER_GIF: Record<SocialSource, number> = {
-    [Source.Twitter]: 15 * 1024 * 1024, // 15MB
-    [Source.Lens]: 15 * 1024 * 1024, // 15MB
-    [Source.Farcaster]: 15 * 1024 * 1024, // 15MB
-};
-
-// https://mask.atlassian.net/browse/FW-2212
-// TODO: Our upload is not strong enough to handle 1GB videos. So we limit 400MB here.
-export const MAX_FILE_SIZE_PER_VIDEO: Record<SocialSource, number> = {
-    [Source.Twitter]: 400 * 1024 * 1024, // 400MB
-    [Source.Lens]: 1024 * 1024 * 1024, // 1GB
-    [Source.Farcaster]: 1024 * 1024 * 1024, // 1GB
-};
 
 // Search Bar
 export const MAX_SEARCH_RECORD_SIZE = 5;
@@ -252,31 +211,6 @@ export const TWITTER_USER_OPTIONS: Partial<UsersV2Params> = {
 };
 
 export const SOLANA_WALLET_CACHE_KEY = 'walletName';
-
-// update profile
-export const MAX_PROFILE_BIO_SIZE: Record<SocialSource, number> = {
-    [Source.Farcaster]: 160,
-    [Source.Lens]: 260,
-    [Source.Twitter]: 160,
-};
-
-export const MAX_PROFILE_DISPLAY_NAME_SIZE: Record<SocialSource, number> = {
-    [Source.Farcaster]: 32,
-    [Source.Lens]: 100,
-    [Source.Twitter]: 50,
-};
-
-export const MAX_PROFILE_LOCATION_SIZE: Record<SocialSource, number> = {
-    [Source.Farcaster]: 0,
-    [Source.Lens]: 100,
-    [Source.Twitter]: 30,
-};
-
-export const MAX_PROFILE_WEBSITE_SIZE: Record<SocialSource, number> = {
-    [Source.Farcaster]: 0,
-    [Source.Lens]: 100,
-    [Source.Twitter]: 100,
-};
 
 // https://support.mirror.xyz/hc/en-us/articles/13729399363220-Platform-fees
 // 0.00069 ETH
