+++ src/services/createFarcasterSchedulePostPayload.ts
@@ -3,7 +3,7 @@ import { toInteger, uniqBy } from 'lodash-es';
 import { toBytes } from 'viem';
 
 import { Source, SourceInURL } from '@/constants/enum.js';
-import { MAX_IMAGE_SIZE_PER_POST } from '@/constants/index.js';
+import { MAX_IMAGE_SIZE_PER_POST } from '@/constants/limitation.js';
 import { readChars } from '@/helpers/chars.js';
 import { getAllMentionsForFarcaster } from '@/helpers/getAllMentionsForFarcaster.js';
 import { getPollFrameUrl } from '@/helpers/getPollFrameUrl.js';
